Continued refactor of replay_engine and processors
