+++ starcraft_predictor/event_processors/__init__.py
@@ -1,8 +1,3 @@
+# flake8: noqa
 from starcraft_predictor.event_processors.base_event_procesor import BaseEventProcessor
 from starcraft_predictor.event_processors.player_stats_event_processor import PlayerStatsEventProcessor
-
-
-all = [
-    "BaseEventProcessor",
-    "PlayerStatsEventProcessor",
-]
